api-monitoring/requirements.txt
simple json
requests
slackclient
pytest
pyYAML
